# api/src/openapi.yaml
openapi: 3.0.2
info:
  version: 0.1.0
  title: NFT Gallery API
  description: API serving the NFT Gallery Platform
...
security:
  - bearerAuth: []
servers:
  - url: '{protocol}://{host}:{port}'
    description: 'Local Mocks (e.g. http://localhost:4010) and Local Dev Server (e.g. http://localhost:3000/v1)'
    variables:
      protocol:
        enum:
          - http
          - https
        default: http
        description: Protocol of the URL.
      host:
        enum:
          - api-dev.domi.house
          - api.domi.house
          - localhost
        default: localhost
        description: Host of the URL
      port:
        enum:
          - ''
          - '4010'
          - '3000'
        default: '3000'
        description: Port of the URL
components:
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Authentication token
paths:
...
  /v1/notification:
    post:
      summary: Create Notification
      operationId: createNotification
      x-eov-operation-handler: NotificationsController
      x-security-scopes:
        - consumer
        - admin
      tags:
        - Notification
      requestBody:
        description: information that is needed for Notification
        required: true
        content:
          application/json:
            schema:
              title: Notification
              type: object
              description: A single Notification received by a user
              x-autogen:
                apipaths: true
                mongo: true
                ownership: true
                security:
                  create:
                    - consumer
                    - admin
                  update:
                    - consumer
                    - admin
                  delete:
                    - consumer
                    - admin
                  retrieve:
                    - consumer
                    - admin
                  retrieves:
                    - consumer
                    - admin
              properties:
                title:
                  type: string
                  description: A possible title (doesn't seem it is needed but better to have it)
                  example: A possible title
                body:
                  type: string
                  description: Text body of the Post
                  example: Something interesting to others
                click_action:
                  type: string
                  description: In App click_action
                  example: reminder/rewiuhfdsl
                type:
                  type: string
                  description: Notification type
                  default: wall
                  enum:
                    - user
                    - asset
                    - bid
                    - chat
                status:
                  type: string
                  description: Notification status
                  default: new
                  enum:
                    - new
                    - hidden
                    - read
                    - deleted
              required:
                - title
                - body
                - status
      responses:
        '200':
          description: Operation Executed
        default:
          description: A standard error object.
          content:
            application/json:
              schema:
                properties:
                  code:
                    type: string
                    x-faker: random.number
                  message:
                    type: string
                    example: Sample error message
                required:
                  - code
                x-tags:
                  - Common
  /v1/notifications:
    parameters:
      - name: limit
        in: query
        description: How many items to return at one time (max 100)
        required: false
        schema:
          type: integer
          format: int32
      - name: skip
        in: query
        description: How many items should be skipped before returnig the required amount
        required: false
        schema:
          type: integer
          format: int32
    get:
      summary: Get all Notifications
      operationId: retrieveNotifications
      x-eov-operation-handler: NotificationsController
      x-security-scopes:
        - consumer
        - admin
      tags:
        - Notification
      parameters:
        - name: title
          in: query
          required: false
          description: The Fileter title
          schema:
            type: string
            example: New Title
        - name: body
          in: query
          required: false
          description: The Fileter body
          schema:
            type: string
            example: Something interesting to others
        - name: click_action
          in: query
          required: false
          description: The Fileter click_action
          schema:
            type: string
            example: reminder&#x2F;rewiuhfdsl
        - name: type
          in: query
          required: false
          description: The Fileter type
          schema:
            type: string
            default: wall
            enum:
              - user
              - asset
              - bid
              - chat
        - name: status
          in: query
          required: false
          description: The Fileter status
          schema:
            type: string
            default: new
            enum:
              - new
              - hidden
              - read
              - deleted
      responses:
        '200':
          description: Operation Executed
        default:
          description: A standard error object.
          content:
            application/json:
              schema:
                properties:
                  code:
                    type: string
                    x-faker: random.number
                  message:
                    type: string
                    example: Sample error message
                required:
                  - code
                x-tags:
                  - Common
  '/v1/notifications/{notificationId}':
    parameters:
      - name: notificationId
        in: path
        required: true
        description: The Id of the Notification
        schema:
          description: Id of the entity
          type: string
          format: not-empty-string
          example: 60758e80d59c357138d01af6
    get:
      summary: Get Notification with Id
      operationId: retrieveNotification
      x-eov-operation-handler: NotificationsController
      x-security-scopes:
        - consumer
        - admin
      tags:
        - Notification
      parameters:
        - name: notificationId
          in: path
          required: true
          description: The Id of the Notification
          schema:
            description: Id of the entity
            type: string
            format: not-empty-string
            example: 60758e80d59c357138d01af6
      responses:
        '200':
          description: Operation Executed
        default:
          description: A standard error object.
          content:
            application/json:
              schema:
                properties:
                  code:
                    type: string
                    x-faker: random.number
                  message:
                    type: string
                    example: Sample error message
                required:
                  - code
                x-tags:
                  - Common
    put:
      summary: Update Notification with Id
      operationId: updateNotification
      x-eov-operation-handler: NotificationsController
      x-security-scopes:
        - consumer
        - admin
      tags:
        - Notification
      parameters:
        - name: notificationId
          in: path
          required: true
          description: The Id of the Notification
          schema:
            description: Id of the entity
            type: string
            format: not-empty-string
            example: 60758e80d59c357138d01af6
      requestBody:
        description: information that is needed for Notification
        required: true
        content:
          application/json:
            schema:
              title: Update Notification with Id
              type: object
              description: An array of fields that we want to change name and value
              properties:
                title:
                  type: string
                  description: A possible title (doesn&amp;#39;t seem it is needed but better to have it)
                  example: A possible title
                body:
                  type: string
                  description: Text body of the Post
                  example: Something interesting to others
                click_action:
                  type: string
                  description: In App click_action
                  example: reminder&amp;#x2F;rewiuhfdsl
                type:
                  type: string
                  description: Notification type
                  default: wall
                  enum:
                    - user
                    - asset
                    - bid
                    - chat
                status:
                  type: string
                  description: Notification status
                  default: new
                  enum:
                    - new
                    - hidden
                    - read
                    - deleted
      responses:
        '200':
          description: Operation Executed
        default:
          description: A standard error object.
          content:
            application/json:
              schema:
                properties:
                  code:
                    type: string
                    x-faker: random.number
                  message:
                    type: string
                    example: Sample error message
                required:
                  - code
                x-tags:
                  - Common
    delete:
      summary: Delete Notification with Id
      operationId: deleteNotification
      x-eov-operation-handler: NotificationsController
      x-security-scopes:
        - consumer
        - admin
      tags:
        - Notification
      parameters:
        - name: notificationId
          in: path
          required: true
          description: The Id of the Notification
          schema:
            description: Id of the entity
            type: string
            format: not-empty-string
            example: 60758e80d59c357138d01af6
      responses:
        '200':
          description: Operation Executed
        default:
          description: A standard error object.
          content:
            application/json:
              schema:
                properties:
                  code:
                    type: string
                    x-faker: random.number
                  message:
                    type: string
                    example: Sample error message
                required:
                  - code
                x-tags:
                  - Common
...